Bump Camel, fcrepo-camel versions
Currently, Camel is at 2.17.0.  An attempt to use the latest
snapshot (containing fix for CAMEL-9462) didn't seem to work due to
validation of the blueprint schema contained within camel-blueprint.
Upgrading to 2.17.0 caused an exception due to read on a closed stream.
Using the jetty producer endpoint (as oposed to http4) seems to be
a workable solution.

https://issues.apache.org/jira/browse/CAMEL-9462