Persist extensions with bnodes as binary
If we load an extension doc that contains blank nodes (e.g. in the case of
embedding ontology statements), persist the extension in the repository as
a binary, allowing straightforward use of reasoners over its contents.
This is a workaround to prevent Fedora from chopping theresource into many
little skolemized bits unuseable to reasoners.