Log error message if resource cennot be reasoned
If an rdf resource does not have a triple that has the resource URI as a
subject, then we rally can't reason over it, nor bind extensions to it.
This is most pertinent to fcr:metadata; we can't bind extensions to
fcr:metadata because it actually describes the binary, not the
fcr:metadata resource itself.