Fix message broker reference in blueprint
Previously, the broker reference was "optional" to accommodate the
scenarios that do not use a broker (tests, etc).  Unfortunately, this
results nondeterministic behaviour when there *is* a broker, depending on
the the order of initialization.  If the broker service is not available
when this bundle starts, then its service reference remains null.