Align with fcrepo-camel-toolbox, #94 (#100)
* Initial cut at fcrepo-camel-toolbox alignment

Note: dependencies are to -SNAPSHOT versions of fcrepo-camel and
fcrepo-camel-toolbox until the next release(s) are out.

* Add integration tests and improve indexing routes

Use an instance of Fuseki to verify sparql update and deletes;  Move to
more idiomatic usage of the new fcrepo-camel-toolbox.  Simplify ITs by
dynamically generating extensions rather than adding relevant objects to
test resources.

* Use release versions of fcrepo-camel-*

* Use named graphs and sparql/update proper

Fully embraces named graph approach to indexing, and refactors the sparql
processors to use application/sparql-update rather than
application/x-www-form-urlencoded for simplicity.

* Add camel unit tests for service doc indexer routes

* Make 410 errors result in ResourceNotFoundException

* Fix message broker reference in blueprint

Previously, the broker reference was "optional" to accommodate the
scenarios that do not use a broker (tests, etc).  Unfortunately, this
results nondeterministic behaviour when there *is* a broker, depending on
the the order of initialization.  If the broker service is not available
when this bundle starts, then its service reference remains null.

* Fix OSGi filtering annotations, and clean up

* Use full URI for extension container

* Handle 410 errors for deleted resources

* Accept activityStream events; cleanup

* Add & test proxy<->noproxy URI mapping methods

* Filter and map message URIs to fedora URIs

* Make sure listener context starts before ITs run

* Dial back default logging level for ITs